Fix new post event is update on redux state
