+++ src/store/eventDetails/reducer.js
@@ -1,5 +1,5 @@
 /* eslint-disable import/no-anonymous-default-export */
-import { EVENT_DETAILS_FETCHED, EVENT_POST_SUCCESS } from "./actions";
+import { EVENT_DETAILS_FETCHED } from "./actions";
 
 const initialState = {};
 
@@ -8,8 +8,8 @@ export default (state = initialState, { type, payload }) => {
     case EVENT_DETAILS_FETCHED:
       return { ...state, ...payload };
     //Update eventDetails/event State
-    case EVENT_POST_SUCCESS:
-      return { ...state, ...payload };
+    // case EVENT_POST_SUCCESS:
+    //   return { ...state, ...payload };
 
     default:
       return state;
